feat: full gameplay — tunnels, self-hit, crush, one-hit death, turn log, player colors
Core physics:
- Trajectory fires from tank body center (x+24) for accurate self-hit alignment
- firstCollision: leftStart + descending flags replace permanent shooter exclusion
  — self-hits allowed once missile passes apex and starts falling
- Tank hit zone widened to full column width (48×30px) for playable precision
- Terrain: right side row-5 blocks removed (cols 14-19) opening ground-floor tunnel

Movement:
- moveTank uses terrainFloorY (scan from current height) + canStand ceiling check
  — tanks traverse tunnels and climb stairs up to 1 block; two-step: staircase then tunnel
- One-hit death: any tank hit ends game immediately

Terrain destruction:
- Block destroyed → tanks fall to next solid floor (terrainFloorY)
- Crush mechanic: falling tank kills tank occupying same landing column/floor
- Tank falls off bottom of map → instant death

UI:
- QML arc uses leftStart pattern matching C++ — arc terminates correctly on return
- P1 yellow (#ffdd00), P2 red (#ff4444); terrain stays green
- AimControl accent color follows active player (yellow/red bars + fire button)
- Turn log strip (RoomPanel): last 8 turns with player, angle, power, outcome
- Debug overlay removed from canvas; internal qDebug logs retained

+++ scorched-earth/src/game_interface.h
@@ -19,7 +19,7 @@ public:
     Q_INVOKABLE virtual int moveTank(int tankId, int direction) = 0;
 
     // Fire shot. Returns JSON: {hit, tankId, removedBlocks, seq, status}
-    Q_INVOKABLE virtual QString processShot(int tankId, float angle, float power) = 0;
+    Q_INVOKABLE virtual QString processShot(int tankId, double angle, double power) = 0;
 
     // 0=ongoing, 1=p1wins, 2=p2wins
     Q_INVOKABLE virtual int gameStatus() = 0;
